Extract trait for shared testing code
