+++ packages/framework/tests/Unit/ColoredBlockquoteShortcodesTest.php
@@ -6,20 +6,15 @@ namespace Hyde\Framework\Testing\Unit;
 
 use Hyde\Markdown\Processing\ColoredBlockquotes;
 use Hyde\Testing\UnitTestCase;
-use Illuminate\Contracts\View\Factory as FactoryContract;
-use Illuminate\Events\Dispatcher;
-use Illuminate\Filesystem\Filesystem;
-use Illuminate\View\Compilers\BladeCompiler;
-use Illuminate\View\Engines\CompilerEngine;
-use Illuminate\View\Engines\EngineResolver;
-use Illuminate\View\Factory;
-use Illuminate\View\FileViewFinder;
+use Hyde\Testing\UsesRealBladeInUnitTests;
 
 /**
  * @covers \Hyde\Markdown\Processing\ColoredBlockquotes
  */
 class ColoredBlockquoteShortcodesTest extends UnitTestCase
 {
+    use UsesRealBladeInUnitTests;
+
     protected static bool $needsKernel = true;
     protected static bool $needsConfig = true;
 
@@ -115,26 +110,5 @@ class ColoredBlockquoteShortcodesTest extends UnitTestCase
                 HTML,
             ],
         ];
-    }
-
-    // Todo: Extract trait for this and MarkdownHeadingRendererUnitTest
-    protected function createRealBladeCompilerEnvironment(): void
-    {
-        $resolver = new EngineResolver();
-        $filesystem = new Filesystem();
-
-        $resolver->register('blade', function () use ($filesystem) {
-            return new CompilerEngine(
-                new BladeCompiler($filesystem, sys_get_temp_dir())
-            );
-        });
-
-        $finder = new FileViewFinder($filesystem, [realpath(__DIR__.'/../../resources/views')]);
-        $finder->addNamespace('hyde', realpath(__DIR__.'/../../resources/views'));
-
-        $view = new Factory($resolver, $finder, new Dispatcher());
-
-        app()->instance('view', $view);
-        app()->instance(FactoryContract::class, $view);
     }
 }
