+++ packages/testing/src/UsesRealBladeInUnitTests.php
@@ -4,7 +4,34 @@ declare(strict_types=1);
 
 namespace Hyde\Testing;
 
+use Illuminate\Contracts\View\Factory as FactoryContract;
+use Illuminate\Events\Dispatcher;
+use Illuminate\Filesystem\Filesystem;
+use Illuminate\View\Compilers\BladeCompiler;
+use Illuminate\View\Engines\CompilerEngine;
+use Illuminate\View\Engines\EngineResolver;
+use Illuminate\View\Factory;
+use Illuminate\View\FileViewFinder;
+
 trait UsesRealBladeInUnitTests
 {
-    //
+    protected function createRealBladeCompilerEnvironment(): void
+    {
+        $resolver = new EngineResolver();
+        $filesystem = new Filesystem();
+
+        $resolver->register('blade', function () use ($filesystem) {
+            return new CompilerEngine(
+                new BladeCompiler($filesystem, sys_get_temp_dir())
+            );
+        });
+
+        $finder = new FileViewFinder($filesystem, [realpath(__DIR__.'/../../resources/views')]);
+        $finder->addNamespace('hyde', realpath(__DIR__.'/../../resources/views'));
+
+        $view = new Factory($resolver, $finder, new Dispatcher());
+
+        app()->instance('view', $view);
+        app()->instance(FactoryContract::class, $view);
+    }
 }
